Update 02-Create a Web App.md
I've been trying to use the image provided in the lab, but the container couldn't be deployed successfully. I researched the issue, and found out that the correct image is "mcr.microsoft.com/azuredocs/aci-helloworld"
I tried creating the webapp with this image, and it was deployed successfully. So, I have modified that property in the creation step.